requirements.txt
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.3
openai==1.57.4
fastapi==0.115.6
uvicorn==0.34.0
